+++ CHANGES.md
@@ -93,6 +93,7 @@
 -   pat autofocus: Implement documented behavior to not focus on prefilled element, if there is another autofocus element which is empty.
 -   pat autofocus: Instead of calling autofocus for each element call it only once.
 -   pat autofocus: Register event handler only once.
+-   pat-checklist: For global de/select buttons, do not change any other checkboxes than the ones the de/select button belongs to.
 
 
 ## 3.0.0-dev - unreleased